requirements.txt
# AI / Orchestration
langchain-groq
langchain-community
langgraph
langsmith

# UI
gradio == 5.42.0

# File parsing
pylatexenc
tavily-python
PyPDF2
docx2txt
mammoth

# HTTP & scraping
requests
beautifulsoup4

# Config management
pydantic-settings
